Remove one errant  / from figures.yaml
I tried using \(c\) to keep the (c) from turning into a copyright symbol, but it didn't work in the figures.yaml file. I realize this one final slash was the wrong way, but I did try it with the correct backslashes. I was able to fix one like this in an md file, but not here.