Issue #117
For objects with 4 views, I've indicated which direction the handle faces and what the label is, as well as some visual cues that informed my choice. Some object views did not include a "Main view" so I did not want to alter that.
Of course I can make any changes as needed.
289: Left Facing, "Main view"
290: Right Facing, "Main view", shows more of the three swirls around neck
292: Has no handle?
295: Left Facing, "Main view"
296: Right Facing, "Main view"
298: Left Facing, "Side B/A", shows the ridge on neck
300: Right Facing, "Side A"
301: Left Facing, "Side B/A"